requirements.txt
absl-py==2.1.0
accelerate==1.12.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
altair==6.0.0
annotated-types==0.7.0
anyio==4.4.0
astroid==3.3.10
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
bcrypt==4.2.1
bitsandbytes==0.48.2
blinker==1.8.2
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
credentials==1.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
dill==0.4.0
distro==1.9.0
edge-tts==7.2.3
elevenlabs==2.26.1
et_xmlfile==2.0.0
eval_type_backport==0.2.0
executing==2.0.1
fal_client==0.10.0
ffmpeg-python==0.2.0
filelock==3.16.1
firebase-admin==6.6.0
flake8==7.2.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Mail==0.10.0
flatbuffers==24.3.25
fonttools==4.51.0
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
gast==0.5.4
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.5
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.11.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.3
huggingface-hub==0.36.0
ibm-cloud-sdk-core==3.23.0
ibm-watson==9.0.0
idna==3.7
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
imbalanced-learn==0.12.3
imblearn==0.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.26.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.2
jupyter_core==5.7.2
keras==3.3.3
kiwisolver==1.4.5
langchain==0.3.1
langchain-core==0.3.7
langchain-openai==0.2.1
langchain-text-splitters==0.3.0
langsmith==0.1.129
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
namex==0.0.8
narwhals==2.13.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.51.0
opencv-python==4.10.0.84
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.11.0
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.2
plotly==6.5.0
pluggy==1.6.0
proglog==0.1.12
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.0.0
psycopg2-binary==2.9.11
pure-eval==0.2.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.12.0
pydantic_core==2.20.1
pydeck==0.9.1
pyflakes==3.3.2
pygame==2.5.2
Pygments==2.18.0
PyJWT==2.10.0
pylint==3.3.7
pyparsing==3.1.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.37.0
regex==2024.5.15
requests==2.32.3
rich==13.9.4
rpds-py==0.30.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.14.0
seaborn==0.13.2
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
square==0.0.3
stack-data==0.6.3
streamlit==1.52.1
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
together==1.3.5
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.57.3
typer==0.13.1
typing-inspection==0.4.2
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
watchdog==6.0.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
yarl==1.13.1
